ai-agent/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
websockets==12.0
langchain==0.1.20
langgraph==0.0.46
langchain-google-genai==1.0.2
google-generativeai>=0.5.0
pydantic==2.10.3
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.24.1
pymongo==4.8.0
redis>=5.0.0

# Optional — required only when USE_FINETUNED_MODEL=true (GPU server)
# torch>=2.2.0
# transformers>=4.40.0
# peft>=0.10.0
# bitsandbytes>=0.43.1
# accelerate>=0.29.3
